app-builder/plugins/requirements.txt
pydantic==2.7.4
psutil==6.1.1
httpx==0.28.1
pandas==2.1.3
linkup-sdk
exa-py
openai
python-dotenv
tavily-python